created custom 404 file
