+++ app/not-found.tsx
@@ -1,0 +1,35 @@
+import Head from "next/head";
+import Link from "next/link";
+
+export default function Custom404() {
+  return (
+    <>
+      <Head>
+        <title>404 - Page Not Found</title>
+      </Head>
+      <section className="relative">
+        <div className="mx-auto max-w-6xl px-4 sm:px-6">
+          <div className="pb-12 pt-32 md:pb-20 md:pt-20">
+            <div className="mx-auto max-w-xl">
+              <div className="mb-10 flex items-center justify-center">
+                <video
+                  autoPlay
+                  loop
+                  muted
+                  playsInline
+                  className="h-auto max-w-full"
+                >
+                  <source
+                    src="/gifs/dukepanshake-404.webm"
+                    type="video/webm"
+                  />
+                  Your browser does not support the video tag.
+                </video>
+              </div>
+            </div>
+          </div>
+        </div>
+      </section>
+    </>
+  );
+}
